+++ src/Oro/Bundle/OrderBundle/Tests/Unit/Entity/OrderShippingTrackingTest.php
@@ -15,11 +15,11 @@ class OrderShippingTrackingTest extends \PHPUnit\Framework\TestCase
         $properties = [
             ['id', 123],
             ['method', 'test_method'],
-            ['number', '1F2B3C4A'],
-            ['order', new Order()]
+            ['number', '1F2B3C4A']
         ];
 
         $orderShippingTracking = new OrderShippingTracking();
         $this->assertPropertyAccessors($orderShippingTracking, $properties);
+        self::assertPropertyCollection($orderShippingTracking, 'orders', new Order());
     }
 }