+++ src/Oro/Bundle/PaymentTermBundle/Form/Extension/SubOrderPaymentTermAclExtension.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Oro\Bundle\PaymentTermBundle\Form\Extension;
+
+use Oro\Bundle\OrderBundle\Form\Type\SubOrderType;
+
+/**
+ * ACL restrictions for SubOrderType form type.
+ */
+class SubOrderPaymentTermAclExtension extends AbstractPaymentTermAclExtension
+{
+    #[\Override]
+    public static function getExtendedTypes(): iterable
+    {
+        return [SubOrderType::class];
+    }
+}